gtk-doc-plugins.mak (-module): Error out when the html build step gives warnings, so they get fixed properly.
Original commit message from CVS:
* gtk-doc-plugins.mak (-module):
Error out when the html build step gives warnings, so they get
fixed properly.

+++ gtk-doc-plugins.mak
@@ -212,7 +212,11 @@ html-build.stamp: sgml.stamp $(DOC_MAIN_SGML_FILE) $(content_files)
 	@for f in $(content_files); do cp $(srcdir)/$$f html; done
 	cp -pr xml html
 	cp ../version.entities html
-	cd html && gtkdoc-mkhtml $(DOC_MODULE) $(DOC_MAIN_SGML_FILE)
+	cd html && gtkdoc-mkhtml $(DOC_MODULE) $(DOC_MAIN_SGML_FILE) \
+	    2>&1 | tee ../html-build.log
+	@if grep "warning:" html-build.log > /dev/null; then \
+		echo "ERROR"; grep "warning:" html-build.log; exit 1; fi
+	@rm html-build.log
 	rm -f html/$(DOC_MAIN_SGML_FILE)
 	rm -rf html/xml
 	rm -f html/version.entities
